Revision: Complete 'Activity 3-4' (ANALYZE, DESIGN! Fruit Ads API)
SSL Database »
    - `fruit`.`image` now supports 512 character length. URL’s
shouldn’t be that long, but links to images seem to be pretty long in
my personal experience. The script will alert the user if the images’
URL was too long, and tell them the record wasn’t added.

fruitads.php »
    - Integrate the old separate content within ‘ads.php’ for display
here, instead.
    - Allow remote image values, as well as local. This is evaluated by
searching the `fruits`.`image` value for a match of the string ‘http’
    - Add ‘image’ field for fruit creation
    - Remove $randomNumber as it creates a bug